Health wip
Added a damage script for damagable hitboxes.
Changed a few thing in the health script.
does not work fully yet

+++ Reusable components/Assets/Scripts/Interfaces.cs
@@ -30,7 +30,7 @@ interface ITrajectory
 
 interface IHealth
 {
-    int HealthInt { get; }
+    float HealthInt { get; }
     void TakeDamage(int damage, bool spriteFlippedX);
     bool knocked { get; }
     //void OnDeath();
